Improve single value selector rendering
